defined order service controller and repository

+++ src/main/java/com/fatichdev/online_food_ordering/model/Order.java
@@ -43,5 +43,5 @@ public class Order {
 
     private int totalItem;
 
-    private double totalPrice;
+    private Long totalPrice;
 }
